Added seperate unit test class for NameSorter class.
